# armory/backend/modules/armory-server/api/http/http.go
// Code generated by fino2. DO NOT EDIT.
// Rerunning fino will overwrite this file.
// Version: v0.1.0
// Version Date:

package http

// This file provides server-side bindings for the HTTP transport.
// It utilizes the transport/http.Server.

import (
	"bytes"
	"compress/gzip"
	"context"
	"encoding/json"
	"fmt"
	"io"
	"io/ioutil"
	"net/http"
	"reflect"
	"strconv"
	"strings"

	"github.com/chaos-io/chaos/logs"
	"github.com/chaos-io/core/go/chaos/core"
	pagination "github.com/chaos-io/gokit/pagination"
	"github.com/chaos-io/gokit/tracing"
	nhttp "github.com/chaos-io/gokit/transport/http"
	"github.com/go-kit/kit/log"
	httptransport "github.com/go-kit/kit/transport/http"
	"github.com/gorilla/mux"
	"github.com/json-iterator/go"
	"github.com/pkg/errors"
	"go.opentelemetry.io/otel/trace"

	// this service api
	"github.com/liankui/OpenBambuAPI/armory/backend/modules/armory-server/api/endpoint"
	pb "github.com/liankui/OpenBambuAPI/armory/go/armory/v1"
)

const contentType = "application/json; charset=utf-8"

var (
	_ = fmt.Sprint
	_ = bytes.Compare
	_ = strconv.Atoi
	_ = httptransport.NewServer
	_ = ioutil.NopCloser
	_ = pb.NewArmoryClient
	_ = io.Copy
	_ = errors.Wrap
)

var cfg *nhttp.Config

func init() {
	cfg = nhttp.NewConfig()
}

// RegisterHttpHandler register a set of endpoints available on predefined paths to the router.
func RegisterHttpHandler(router *mux.Router, endpoints endpoint.Endpoints, tracer trace.Tracer, logger log.Logger) {
	serverOptions := []httptransport.ServerOption{
		httptransport.ServerBefore(headersToContext, queryToContext),
		httptransport.ServerErrorEncoder(errorEncoder),
		httptransport.ServerErrorLogger(logger),
		httptransport.ServerAfter(httptransport.SetContentType(contentType)),
	}

	addTracerOption := func(methodName string) []httptransport.ServerOption {
		if tracer != nil {
			return append(serverOptions, httptransport.ServerBefore(tracing.HTTPToContext(tracer, methodName)))
		}
		return serverOptions
	}

	router.Methods("POST").Path("/v1/user-service/user/login").Handler(
		httptransport.NewServer(
			endpoints.LoginEndpoint,
			DecodeHTTPLoginRequest,
			EncodeHTTPGenericResponse,
			addTracerOption("Login")...,
		))

	router.Methods("POST").Path("/v1/user-service/user/refreshtoken").Handler(
		httptransport.NewServer(
			endpoints.RefreshTokenEndpoint,
			DecodeHTTPRefreshTokenRequest,
			EncodeHTTPGenericResponse,
			addTracerOption("RefreshToken")...,
		))

	router.Methods("GET").Path("/v1/user-service/my/messages").Handler(
		httptransport.NewServer(
			endpoints.ListMyMessagesEndpoint,
			DecodeHTTPListMyMessagesRequest,
			EncodeHTTPGenericResponse,
			addTracerOption("ListMyMessages")...,
		))

	router.Methods("GET").Path("/v1/user-service/my/tasks").Handler(
		httptransport.NewServer(
			endpoints.ListMyTasksEndpoint,
			DecodeHTTPListMyTasksRequest,
			EncodeHTTPGenericResponse,
			addTracerOption("ListMyTasks")...,
		))

	router.Methods("GET").Path("/v1/iot-service/api/slicer/resource").Handler(
		httptransport.NewServer(
			endpoints.GetSlicerResourcesEndpoint,
			DecodeHTTPGetSlicerResourcesRequest,
			EncodeHTTPGenericResponse,
			addTracerOption("GetSlicerResources")...,
		))

	router.Methods("GET").Path("/v1/iot-service/api/user/bind").Handler(
		httptransport.NewServer(
			endpoints.ListUserDevicesEndpoint,
			DecodeHTTPListUserDevicesRequest,
			EncodeHTTPGenericResponse,
			addTracerOption("ListUserDevices")...,
		))

	router.Methods("GET").Path("/v1/iot-service/api/user/print").Handler(
		httptransport.NewServer(
			endpoints.GetPrintStatusEndpoint,
			DecodeHTTPGetPrintStatusRequest,
			EncodeHTTPGenericResponse,
			addTracerOption("GetPrintStatus")...,
		))

	router.Methods("POST").Path("/v1/iot-service/api/user/ttcode").Handler(
		httptransport.NewServer(
			endpoints.GetTtcodeEndpoint,
			DecodeHTTPGetTtcodeRequest,
			EncodeHTTPGenericResponse,
			addTracerOption("GetTTCode")...,
		))
}

// ErrorEncoder writes the error to the ResponseWriter, by default a content
// type of application/json, a body of json with key "error" and the value
// error.Error(), and a status code of 500. If the error implements Headerer,
// the provided headers will be applied to the response. If the error
// implements json.Marshaler, and the marshaling succeeds, the JSON encoded
// form of the error will be used. If the error implements StatusCoder, the
// provided StatusCode will be used instead of 500.
func errorEncoder(ctx context.Context, err error, w http.ResponseWriter) {
	e, ok := err.(*core.Error)
	if !ok {
		e = core.NewErrorFrom(500, err.Error())
	}

	w.Header().Set("Content-Type", contentType)
	if headerer, ok := err.(httptransport.Headerer); ok {
		for k := range headerer.Headers() {
			w.Header().Set(k, headerer.Headers().Get(k))
		}
	}

	var body []byte
	code := http.StatusInternalServerError

	if enveloped := nhttp.IsEnvelopeStyle(ctx, cfg.GetStyle()); enveloped {
		envelope := &nhttp.EnvelopedResponse{}
		envelope.Error = e

		if cfg.GetEnvelop().MappingCode {
			if sc, ok := err.(httptransport.StatusCoder); ok {
				code = sc.StatusCode()
			}
		} else {
			code = http.StatusOK
		}

		var response interface{}
		if cfg.GetEnvelop().ErrorWrapped {
			response = envelope.ToErrorWrapped()
		} else {
			response = envelope
		}

		jsonBody, marshalErr := jsoniter.ConfigFastest.Marshal(response)
		if marshalErr != nil {
			logs.Warnw("failed to marshal the error response to json", "error", marshalErr)
		} else {
			body = jsonBody
		}
	} else {
		if sc, ok := err.(httptransport.StatusCoder); ok {
			code = sc.StatusCode()
		}
		if marshaler, ok := err.(json.Marshaler); ok {
			if jsonBody, marshalErr := marshaler.MarshalJSON(); marshalErr == nil {
				body = jsonBody
			}
		}

		if jsonBody, marshalErr := jsoniter.ConfigFastest.Marshal(e); marshalErr == nil {
			body = jsonBody
		}
	}

	w.WriteHeader(code)
	w.Write(body)
}

// Server Decode

// DecodeHTTPLoginRequest is a transport/http.DecodeRequestFunc that
// decodes a JSON-encoded login request from the HTTP request
// body. Primarily useful in a server.
func DecodeHTTPLoginRequest(_ context.Context, r *http.Request) (interface{}, error) {
	var req pb.LoginRequest

	// to support gzip input
	var reader io.ReadCloser
	var err error
	switch r.Header.Get("Content-Encoding") {
	case "gzip":
		reader, err = gzip.NewReader(r.Body)
		defer reader.Close()
		if err != nil {
			return nil, nhttp.WrapError(err, 400, "failed to read the gzip content")
		}
	default:
		reader = r.Body
	}

	buf, err := io.ReadAll(reader)
	if err != nil {
		return nil, nhttp.WrapError(err, 400, "cannot read body of http request")
	}
	if len(buf) > 0 {
		if err = jsoniter.ConfigFastest.Unmarshal(buf, &req); err != nil {
			const size = 8196
			if len(buf) > size {
				buf = buf[:size]
			}
			return nil, nhttp.WrapError(err,
				http.StatusBadRequest,
				fmt.Sprintf("request body '%s': cannot parse non-json request body", buf),
			)
		}
	}

	pathParams := mux.Vars(r)
	_ = pathParams

	queryParams := core.NewUrlQueryFrom(r.URL.Query())
	_ = queryParams

	parsedQueryParams := make(map[string]bool)
	_ = parsedQueryParams

	err = unmarshalQueryParam(queryParams, &req.Account, "account")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the account query parameter")
	}

	err = unmarshalQueryParam(queryParams, &req.Password, "password")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the password query parameter")
	}

	err = unmarshalQueryParam(queryParams, &req.Code, "code")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the code query parameter")
	}

	return &req, nil
}

// DecodeHTTPRefreshTokenRequest is a transport/http.DecodeRequestFunc that
// decodes a JSON-encoded refreshtoken request from the HTTP request
// body. Primarily useful in a server.
func DecodeHTTPRefreshTokenRequest(_ context.Context, r *http.Request) (interface{}, error) {
	var req pb.RefreshTokenRequest

	// to support gzip input
	var reader io.ReadCloser
	var err error
	switch r.Header.Get("Content-Encoding") {
	case "gzip":
		reader, err = gzip.NewReader(r.Body)
		defer reader.Close()
		if err != nil {
			return nil, nhttp.WrapError(err, 400, "failed to read the gzip content")
		}
	default:
		reader = r.Body
	}

	buf, err := io.ReadAll(reader)
	if err != nil {
		return nil, nhttp.WrapError(err, 400, "cannot read body of http request")
	}
	if len(buf) > 0 {
		if err = jsoniter.ConfigFastest.Unmarshal(buf, &req); err != nil {
			const size = 8196
			if len(buf) > size {
				buf = buf[:size]
			}
			return nil, nhttp.WrapError(err,
				http.StatusBadRequest,
				fmt.Sprintf("request body '%s': cannot parse non-json request body", buf),
			)
		}
	}

	pathParams := mux.Vars(r)
	_ = pathParams

	queryParams := core.NewUrlQueryFrom(r.URL.Query())
	_ = queryParams

	parsedQueryParams := make(map[string]bool)
	_ = parsedQueryParams

	err = unmarshalQueryParam(queryParams, &req.RefreshToken, "refresh_token")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the refresh_token query parameter")
	}

	return &req, nil
}

// DecodeHTTPListMyMessagesRequest is a transport/http.DecodeRequestFunc that
// decodes a JSON-encoded listmymessages request from the HTTP request
// body. Primarily useful in a server.
func DecodeHTTPListMyMessagesRequest(_ context.Context, r *http.Request) (interface{}, error) {
	var req pb.ListMyMessagesRequest

	// to support gzip input
	var reader io.ReadCloser
	var err error
	switch r.Header.Get("Content-Encoding") {
	case "gzip":
		reader, err = gzip.NewReader(r.Body)
		defer reader.Close()
		if err != nil {
			return nil, nhttp.WrapError(err, 400, "failed to read the gzip content")
		}
	default:
		reader = r.Body
	}

	buf, err := io.ReadAll(reader)
	if err != nil {
		return nil, nhttp.WrapError(err, 400, "cannot read body of http request")
	}
	if len(buf) > 0 {
		if err = jsoniter.ConfigFastest.Unmarshal(buf, &req); err != nil {
			const size = 8196
			if len(buf) > size {
				buf = buf[:size]
			}
			return nil, nhttp.WrapError(err,
				http.StatusBadRequest,
				fmt.Sprintf("request body '%s': cannot parse non-json request body", buf),
			)
		}
	}

	pathParams := mux.Vars(r)
	_ = pathParams

	queryParams := core.NewUrlQueryFrom(r.URL.Query())
	_ = queryParams

	parsedQueryParams := make(map[string]bool)
	_ = parsedQueryParams

	err = unmarshalQueryParam(queryParams, &req.Type, "type")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the type query parameter")
	}

	err = unmarshalQueryParam(queryParams, &req.After, "after")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the after query parameter")
	}

	err = unmarshalQueryParam(queryParams, &req.Limit, "limit")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the limit query parameter")
	}

	return &req, nil
}

// DecodeHTTPListMyTasksRequest is a transport/http.DecodeRequestFunc that
// decodes a JSON-encoded listmytasks request from the HTTP request
// body. Primarily useful in a server.
func DecodeHTTPListMyTasksRequest(_ context.Context, r *http.Request) (interface{}, error) {
	var req pb.ListMyTasksRequest

	// to support gzip input
	var reader io.ReadCloser
	var err error
	switch r.Header.Get("Content-Encoding") {
	case "gzip":
		reader, err = gzip.NewReader(r.Body)
		defer reader.Close()
		if err != nil {
			return nil, nhttp.WrapError(err, 400, "failed to read the gzip content")
		}
	default:
		reader = r.Body
	}

	buf, err := io.ReadAll(reader)
	if err != nil {
		return nil, nhttp.WrapError(err, 400, "cannot read body of http request")
	}
	if len(buf) > 0 {
		if err = jsoniter.ConfigFastest.Unmarshal(buf, &req); err != nil {
			const size = 8196
			if len(buf) > size {
				buf = buf[:size]
			}
			return nil, nhttp.WrapError(err,
				http.StatusBadRequest,
				fmt.Sprintf("request body '%s': cannot parse non-json request body", buf),
			)
		}
	}

	pathParams := mux.Vars(r)
	_ = pathParams

	queryParams := core.NewUrlQueryFrom(r.URL.Query())
	_ = queryParams

	parsedQueryParams := make(map[string]bool)
	_ = parsedQueryParams

	err = unmarshalQueryParam(queryParams, &req.DeviceId, "device_id")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the device_id query parameter")
	}

	err = unmarshalQueryParam(queryParams, &req.After, "after")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the after query parameter")
	}

	err = unmarshalQueryParam(queryParams, &req.Limit, "limit")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the limit query parameter")
	}

	return &req, nil
}

// DecodeHTTPGetSlicerResourcesRequest is a transport/http.DecodeRequestFunc that
// decodes a JSON-encoded getslicerresources request from the HTTP request
// body. Primarily useful in a server.
func DecodeHTTPGetSlicerResourcesRequest(_ context.Context, r *http.Request) (interface{}, error) {
	var req pb.GetSlicerResourcesRequest

	// to support gzip input
	var reader io.ReadCloser
	var err error
	switch r.Header.Get("Content-Encoding") {
	case "gzip":
		reader, err = gzip.NewReader(r.Body)
		defer reader.Close()
		if err != nil {
			return nil, nhttp.WrapError(err, 400, "failed to read the gzip content")
		}
	default:
		reader = r.Body
	}

	buf, err := io.ReadAll(reader)
	if err != nil {
		return nil, nhttp.WrapError(err, 400, "cannot read body of http request")
	}
	if len(buf) > 0 {
		if err = jsoniter.ConfigFastest.Unmarshal(buf, &req); err != nil {
			const size = 8196
			if len(buf) > size {
				buf = buf[:size]
			}
			return nil, nhttp.WrapError(err,
				http.StatusBadRequest,
				fmt.Sprintf("request body '%s': cannot parse non-json request body", buf),
			)
		}
	}

	pathParams := mux.Vars(r)
	_ = pathParams

	queryParams := core.NewUrlQueryFrom(r.URL.Query())
	_ = queryParams

	parsedQueryParams := make(map[string]bool)
	_ = parsedQueryParams

	err = unmarshalQueryParam(queryParams, &req.Type, "type")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the type query parameter")
	}

	return &req, nil
}

// DecodeHTTPListUserDevicesRequest is a transport/http.DecodeRequestFunc that
// decodes a JSON-encoded listuserdevices request from the HTTP request
// body. Primarily useful in a server.
func DecodeHTTPListUserDevicesRequest(_ context.Context, r *http.Request) (interface{}, error) {
	var req pb.ListUserDevicesRequest

	// to support gzip input
	var reader io.ReadCloser
	var err error
	switch r.Header.Get("Content-Encoding") {
	case "gzip":
		reader, err = gzip.NewReader(r.Body)
		defer reader.Close()
		if err != nil {
			return nil, nhttp.WrapError(err, 400, "failed to read the gzip content")
		}
	default:
		reader = r.Body
	}

	buf, err := io.ReadAll(reader)
	if err != nil {
		return nil, nhttp.WrapError(err, 400, "cannot read body of http request")
	}
	if len(buf) > 0 {
		if err = jsoniter.ConfigFastest.Unmarshal(buf, &req); err != nil {
			const size = 8196
			if len(buf) > size {
				buf = buf[:size]
			}
			return nil, nhttp.WrapError(err,
				http.StatusBadRequest,
				fmt.Sprintf("request body '%s': cannot parse non-json request body", buf),
			)
		}
	}

	pathParams := mux.Vars(r)
	_ = pathParams

	queryParams := core.NewUrlQueryFrom(r.URL.Query())
	_ = queryParams

	parsedQueryParams := make(map[string]bool)
	_ = parsedQueryParams
	return &req, nil
}

// DecodeHTTPGetPrintStatusRequest is a transport/http.DecodeRequestFunc that
// decodes a JSON-encoded getprintstatus request from the HTTP request
// body. Primarily useful in a server.
func DecodeHTTPGetPrintStatusRequest(_ context.Context, r *http.Request) (interface{}, error) {
	var req pb.GetPrintStatusRequest

	// to support gzip input
	var reader io.ReadCloser
	var err error
	switch r.Header.Get("Content-Encoding") {
	case "gzip":
		reader, err = gzip.NewReader(r.Body)
		defer reader.Close()
		if err != nil {
			return nil, nhttp.WrapError(err, 400, "failed to read the gzip content")
		}
	default:
		reader = r.Body
	}

	buf, err := io.ReadAll(reader)
	if err != nil {
		return nil, nhttp.WrapError(err, 400, "cannot read body of http request")
	}
	if len(buf) > 0 {
		if err = jsoniter.ConfigFastest.Unmarshal(buf, &req); err != nil {
			const size = 8196
			if len(buf) > size {
				buf = buf[:size]
			}
			return nil, nhttp.WrapError(err,
				http.StatusBadRequest,
				fmt.Sprintf("request body '%s': cannot parse non-json request body", buf),
			)
		}
	}

	pathParams := mux.Vars(r)
	_ = pathParams

	queryParams := core.NewUrlQueryFrom(r.URL.Query())
	_ = queryParams

	parsedQueryParams := make(map[string]bool)
	_ = parsedQueryParams

	err = unmarshalQueryParam(queryParams, &req.Force, "force")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the force query parameter")
	}

	return &req, nil
}

// DecodeHTTPGetTtcodeRequest is a transport/http.DecodeRequestFunc that
// decodes a JSON-encoded getttcode request from the HTTP request
// body. Primarily useful in a server.
func DecodeHTTPGetTtcodeRequest(_ context.Context, r *http.Request) (interface{}, error) {
	var req pb.GetTtcodeRequest

	// to support gzip input
	var reader io.ReadCloser
	var err error
	switch r.Header.Get("Content-Encoding") {
	case "gzip":
		reader, err = gzip.NewReader(r.Body)
		defer reader.Close()
		if err != nil {
			return nil, nhttp.WrapError(err, 400, "failed to read the gzip content")
		}
	default:
		reader = r.Body
	}

	buf, err := io.ReadAll(reader)
	if err != nil {
		return nil, nhttp.WrapError(err, 400, "cannot read body of http request")
	}
	if len(buf) > 0 {
		if err = jsoniter.ConfigFastest.Unmarshal(buf, &req); err != nil {
			const size = 8196
			if len(buf) > size {
				buf = buf[:size]
			}
			return nil, nhttp.WrapError(err,
				http.StatusBadRequest,
				fmt.Sprintf("request body '%s': cannot parse non-json request body", buf),
			)
		}
	}

	pathParams := mux.Vars(r)
	_ = pathParams

	queryParams := core.NewUrlQueryFrom(r.URL.Query())
	_ = queryParams

	parsedQueryParams := make(map[string]bool)
	_ = parsedQueryParams

	err = unmarshalQueryParam(queryParams, &req.DevId, "dev_id")
	if err != nil && !core.IsNotFoundError(err) {
		return nil, nhttp.WrapError(err, 400, "cannot unmarshal the dev_id query parameter")
	}

	return &req, nil
}

// EncodeHTTPGenericResponse is a transport/http.EncodeResponseFunc that encodes
// the response as JSON to the response writer. Primarily useful in a server.
func EncodeHTTPGenericResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
	if writer, ok := response.(nhttp.ResponseWriter); ok {
		return writer.WriteHttpResponse(ctx, w)
	}

	if reflect.ValueOf(response).IsNil() {
		response = nil
	}
	if _, ok := response.(*core.Null); ok {
		response = nil
	}

	enveloped := nhttp.IsEnvelopeStyle(ctx, cfg.GetStyle())
	if enveloped {
		code := core.NewErrorCode(200)
		message := "OK"
		if sc := cfg.GetEnvelop().SuccessCode; len(sc) > 0 {
			if c, err := core.ParseErrorCode(sc); err != nil {
				logs.Warnw("failed to parse the user setting success code, will use \"200\" indeed.", "code", sc, "error", err)
			} else {
				code = c
			}
		}
		if msg := cfg.GetEnvelop().SuccessMessage; len(msg) > 0 {
			message = msg
		}

		totalCount := int32(0)
		nextPageToken := ""
		if p, ok := response.(pagination.Paginater); ok {
			totalCount = p.GetTotalCount()
			nextPageToken = p.GetNextPageToken()
		}

		response = &nhttp.EnvelopedResponse{
			Error: &core.Error{
				Code:    code,
				Message: message,
			},
			TotalCount:    totalCount,
			NextPageToken: nextPageToken,
			Data:          response,
		}
	}

	if response == nil {
		return nil
	}

	if p, ok := response.(pagination.Paginater); ok && !enveloped {
		total := p.GetTotalCount()
		if total > 0 {
			w.Header().Set("X-Total-Count", strconv.Itoa(int(total)))
		}

		next := p.GetNextPageToken()
		if len(next) > 0 {
			path, _ := ctx.Value("http-request-path").(string)
			if len(path) == 0 {
				path = "/?next-page-token=" + next
			} else {
				url, _ := core.ParseUrl(path)
				url.Query.Add("next-page-token", next)
				path = url.Format()
			}
			w.Header().Set("Link", fmt.Sprintf("<%s>; rel=\"next\"", path))
		}
	}

	return nhttp.NewResponseJsonWriter(response).WriteHttpResponse(ctx, w)
}

// Helper functions

func headersToContext(ctx context.Context, r *http.Request) context.Context {
	for k, _ := range r.Header {
		// The key is added both in http format (k) which has had
		// http.CanonicalHeaderKey called on it in transport as well as the
		// strings.ToLower which is the grpc metadata format of the key so
		// that it can be accessed in either format
		ctx = context.WithValue(ctx, k, r.Header.Get(k))
		ctx = context.WithValue(ctx, strings.ToLower(k), r.Header.Get(k))
	}

	// add the access key to context
	accessKey := r.URL.Query().Get("access_key")
	if len(accessKey) > 0 {
		ctx = context.WithValue(ctx, "access_key", accessKey)
	}

	// Tune specific change.
	// also add the request url
	ctx = context.WithValue(ctx, "http-request-path", r.URL.Path)
	ctx = context.WithValue(ctx, "transport", "HTTPJSON")

	return ctx
}

func queryToContext(ctx context.Context, r *http.Request) context.Context {
	check := func(values []string) bool {
		for _, value := range values {
			if value == "true" {
				return true
			}
		}
		return false
	}

	for key, values := range r.URL.Query() {
		switch key {
		case "envelope":
			if check(values) {
				ctx = context.WithValue(ctx, "envelope", true)
			}
		}
	}

	return ctx
}

func unmarshalPathParam(params map[string]string, value interface{}, keys ...string) error {
	val := ""
	key := ""
	for _, key = range keys {
		if val, _ = params[key]; len(val) > 0 {
			break
		}
	}
	if len(val) == 0 {
		return core.NewNotFoundError(key)
	}

	return core.UnmarshalParam(val, value)
}

func unmarshalQueryParam(params *core.Query, value interface{}, keys ...string) error {
	key := ""
	for _, k := range keys {
		if params.Has(k) {
			key = k
		}
	}
	if len(key) == 0 {
		return core.NewNotFoundError(key)
	}

	return params.Unmarshal(key, value)
}