# armory/openapi/armory/v1/armory.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "armory/v1/armory.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/iot-service/api/slicer/resource": {
      "get": {
        "operationId": "Armory_GetSlicerResources",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetSlicerResourcesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Armory"
        ]
      }
    },
    "/v1/iot-service/api/user/bind": {
      "get": {
        "operationId": "Armory_ListUserDevices",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListUserDevicesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Armory"
        ]
      }
    },
    "/v1/iot-service/api/user/print": {
      "get": {
        "operationId": "Armory_GetPrintStatus",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetPrintStatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "Armory"
        ]
      }
    },
    "/v1/iot-service/api/user/ttcode": {
      "post": {
        "operationId": "Armory_GetTTCode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetTTCodeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GetTTCodeRequest"
            }
          }
        ],
        "tags": [
          "Armory"
        ]
      }
    },
    "/v1/user-service/my/messages": {
      "get": {
        "operationId": "Armory_ListMyMessages",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMyMessagesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Armory"
        ]
      }
    },
    "/v1/user-service/my/tasks": {
      "get": {
        "operationId": "Armory_ListMyTasks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMyTasksResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "deviceId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Armory"
        ]
      }
    },
    "/v1/user-service/user/login": {
      "post": {
        "operationId": "Armory_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LoginRequest"
            }
          }
        ],
        "tags": [
          "Armory"
        ]
      }
    },
    "/v1/user-service/user/refreshtoken": {
      "post": {
        "operationId": "Armory_RefreshToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RefreshTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RefreshTokenRequest"
            }
          }
        ],
        "tags": [
          "Armory"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1BaseResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "v1Device": {
      "type": "object",
      "properties": {
        "devId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "online": {
          "type": "boolean"
        },
        "devModelName": {
          "type": "string"
        },
        "devProductName": {
          "type": "string"
        }
      }
    },
    "v1GetPrintStatusResponse": {
      "type": "object",
      "properties": {
        "base": {
          "$ref": "#/definitions/v1BaseResponse"
        },
        "devices": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1PrintDeviceStatus"
          }
        }
      }
    },
    "v1GetSlicerResourcesResponse": {
      "type": "object",
      "properties": {
        "base": {
          "$ref": "#/definitions/v1BaseResponse"
        },
        "software": {
          "$ref": "#/definitions/v1Software"
        },
        "resources": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Resource"
          }
        }
      }
    },
    "v1GetTTCodeRequest": {
      "type": "object",
      "properties": {
        "devId": {
          "type": "string"
        }
      }
    },
    "v1GetTTCodeResponse": {
      "type": "object",
      "properties": {
        "base": {
          "$ref": "#/definitions/v1BaseResponse"
        },
        "ttcode": {
          "type": "string"
        },
        "passwd": {
          "type": "string"
        },
        "authkey": {
          "type": "string"
        }
      }
    },
    "v1ListMyMessagesResponse": {
      "type": "object",
      "properties": {
        "hits": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1UserMessage"
          }
        }
      }
    },
    "v1ListMyTasksResponse": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "hits": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1PrintTask"
          }
        }
      }
    },
    "v1ListUserDevicesResponse": {
      "type": "object",
      "properties": {
        "base": {
          "$ref": "#/definitions/v1BaseResponse"
        },
        "devices": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Device"
          }
        }
      }
    },
    "v1LoginRequest": {
      "type": "object",
      "properties": {
        "account": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      }
    },
    "v1LoginResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "loginType": {
          "type": "string"
        },
        "expiresIn": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "v1PrintDeviceStatus": {
      "type": "object",
      "properties": {
        "devId": {
          "type": "string"
        },
        "devName": {
          "type": "string"
        },
        "devOnline": {
          "type": "boolean"
        },
        "progress": {
          "type": "integer",
          "format": "int32"
        },
        "taskId": {
          "type": "string"
        }
      }
    },
...
    "v1RefreshTokenRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string"
        }
      }
    },
    "v1RefreshTokenResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "expiresIn": {
          "type": "string",
          "format": "int64"
        },
        "refreshExpiresIn": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1Resource": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "forceUpdate": {
          "type": "boolean"
        }
      }
    },
    "v1Software": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "forceUpdate": {
          "type": "boolean"
        }
      }
    },
    "v1TaskMessage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "cover": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "deviceId": {
          "type": "string"
        }
      }
    },
    "v1UserMessage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "type": {
          "type": "integer",
          "format": "int32"
        },
        "taskMessage": {
          "$ref": "#/definitions/v1TaskMessage"
        },
        "from": {
          "$ref": "#/definitions/v1MessageUser"
        },
        "createTime": {
          "type": "string"
        }
      }
    }
  }
}